requirements-dev.txt
absolufy-imports
astroid
numpydoc
pydocstyle
pylint
pytest
pytest-cov
black ; python_version >= '3.6'
typing_copilot ; python_version >= '3.7'
mypy<0.800
